Add wthSuspense in body
