+++ src/utils/index.js
@@ -1,14 +1,16 @@
 import noop from './noop';
 import getTitle from './getTitle';
 import getAllData from './getAllData.js';
+import withSuspense from './withSuspense';
 import transformData from './transformData';
 import { loadState, saveState } from './persist';
 
 export {
   noop,
   getTitle,
+  loadState,
   saveState,
   getAllData,
-  loadState,
+  withSuspense,
   transformData,
 }
